feat(light/fan): functionality for fan and light
